Function smoldot_js::ffi::bindings::add_chain

#[no_mangle]
pub extern "C" fn add_chain(
    chain_spec_pointer: u32, 
    chain_spec_len: u32, 
    json_rpc_running: u32, 
    potential_relay_chains_ptr: u32, 
    potential_relay_chains_len: u32
) -> u32

Adds a chain to the client. The client will try to stay connected and synchronize this chain.

Use alloc to allocate a buffer for the spec of the chain that needs to be started. Write the chain spec in this buffer as UTF-8. Then, pass the pointer and length (in bytes) as parameter to this function.

Similarly, use alloc to allocate a buffer containing a list of 32-bits-little-endian chain ids. Pass the pointer and number of chain ids (not length in bytes of the buffer) to this function. If the chain specification refer to a parachain, these chain ids are the ones that will be looked up to find the corresponding relay chain.

These two buffers must have been allocated with alloc. They are freed when this function is called, even if an error code is returned.

If json_rpc_running is 0, then no JSON-RPC service will be started and all JSON-RPC requests targeting this chain will return an error. This can be used to save up resources.

Returns:

(2^32-1) if an error happened.

Any other value to indicate success. This value is an identifier that will be used to refer to this chain.