Function smoldot_js::ffi::bindings::connection_new

pub unsafe extern "C" fn connection_new(
    id: u32, 
    addr_ptr: u32, 
    addr_len: u32, 
    error_ptr_ptr: u32
) -> u32

Must initialize a new connection that tries to connect to the given multiaddress.

The multiaddress is a UTF-8 string found in the WebAssembly memory at offset addr_ptr and with addr_len bytes. The string is a multiaddres such as /ip4/1.2.3.4/tcp/5/ws.

The id parameter is an identifier for this connection, as chosen by the Rust code. It must be passed on every interaction with this connection.

Returns 0 to indicate success, or 1 to indicate that an error happened. If an error is returned, the id doesn’t correspond to anything.

Note: If you implement this function using for example new WebSocket(), please keep in mind that exceptions should be caught and turned into an error code.

The error_ptr_ptr parameter should be treated as a pointer to two consecutive little-endian 32-bits unsigned numbers. If an error happened, call alloc to allocate memory, write a UTF-8 error message in that given location, then write that location at the location indicated by error_ptr_ptr and the length of that string at the location error_ptr_ptr + 4. The buffer is then de-allocated by the client. If no error happens, nothing should be written to error_ptr_ptr.

At any time, a connection can be in one of the three following states:

Opening (initial state)

Open

Closed

When in the Opening or Open state, the connection can transition to the Closed state if the remote closes the connection or refuses the connection altogether. When that happens, connection_closed must be called. Once in the Closed state, the connection cannot transition back to another state.

Initially in the Opening state, the connection can transition to the Open state if the remote accepts the connection. When that happens, connection_open must be called.

When in the Open state, the connection can receive messages. When a message is received, alloc must be called in order to allocate memory for this message, then connection_message must be called with the pointer returned by alloc.