Function smoldot_js::ffi::bindings::json_rpc_send

#[no_mangle]pub extern "C" fn json_rpc_send(
    text_ptr: u32, 
    text_len: u32, 
    chain_index: u32, 
    user_data: u32
)

Emit a JSON-RPC request. If the initialization (see init) hasn’t been started or hasn’t finished yet, the request will still be queued.

A buffer containing a UTF-8 JSON-RPC request must be passed as parameter. The format of the JSON-RPC requests is described in the standard JSON-RPC 2.0 specifications. A pub-sub extension is supported.

The buffer passed as parameter must have been allocated with alloc. It is freed when this function is called.

Additionally, an arbitrary value is also passed as a parameter. This value will later be provided back in json_rpc_respond. It can also be passed to json_rpc_unsubscribe_all.

Responses and subscriptions notifications are sent back using json_rpc_respond.